% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row}
\alias{add_row}
\title{insert a row into a table}
\usage{
add_row(conn, tbl_name, .dat, verbose = FALSE)
}
\arguments{
\item{conn}{the `DBI` connection}

\item{tbl_name}{the name of the table in the database}

\item{.dat}{a named list where the name is the database table field name and
the each value is of length 1 and is the value for the field}
}
\value{
the number of rows affected or an error
}
\description{
insert a row into a table
}
\examples{
con <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

DBI::dbWriteTable(
  con,
  name = "test",
  value = data.frame(
    id = 1:2,
    data_col = rep("hi", times = 2)
  )
)

test_dat <- list(
  id = 3,
  data_col = "hello"
)

add_row(con, "test", .dat = test_dat)

DBI::dbDisconnect(con)


}
