% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_app.R
\name{bundle_app}
\alias{bundle_app}
\title{bundle_app}
\usage{
bundle_app(path = ".")
}
\arguments{
\item{path}{to the the root directory of the app. defaults to the working directory}
}
\description{
creates a tar ball for the app in "packrat/bundles/" using `packrat`
}
