% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{contact_ui}
\alias{contact_ui}
\title{contact_ui}
\usage{
contact_ui(contacts = list(list(name = "Elon Musk", phone = "555-555-5555",
  email = "elon.musk@tychobra.com")))
}
\arguments{
\item{contacts}{}
}
\description{
ui for contact dropdown button in shinydashboard
}
