% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Database connection module}
\usage{
db_connect(db_config, db_driver = RPostgres::Postgres())
}
\arguments{
\item{db_config}{database configuration}

\item{db_driver}{database driver}
}
\value{
a DBIConnect object
}
\description{
connect to the db
}
