% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_pool}
\alias{db_pool}
\title{Database connection pool module}
\usage{
db_pool(db_config, db_driver = RPostgres::Postgres())
}
\arguments{
\item{db_config}{database configuration}

\item{db_driver}{database driver}
}
\value{
a database pool
}
\description{
create a database connection pool
}
