% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_by_id.R
\name{delete_by_id}
\alias{delete_by_id}
\title{delete_by_id}
\usage{
delete_by_id(conn, tbl_name, id)
}
\arguments{
\item{conn}{the DBI connection}

\item{tbl_name}{the name of the table}

\item{id}{the row id}
}
\value{
the number of rows affected or an error. This value should be 1, 0, or the error
}
\description{
Delete a row from a table by the row id
}
\examples{
con <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

DBI::dbWriteTable(
  con,
  name = "test",
  value = data.frame(
    id = 1:2,
    data_col = rep("hi", times = 2)
  )
)

delete_by_id(con, "test", id = 1)

DBI::dbDisconnect(con)

}
