% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_buttons_options.R
\name{dt_buttons_options}
\alias{dt_buttons_options}
\title{options for DT buttons}
\usage{
dt_buttons_options(filename = function() paste0("file-", Sys.Date(),
  ".xlsx"), cols = NULL)
}
\arguments{
\item{filename}{function that returns the filename for the downloaded file}

\item{cols}{columns to download.  Passed to exportOptions}
}
\description{
options for DT buttons
}
