% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_callback.R
\name{dt_callback}
\alias{dt_callback}
\title{DT Responsive Table Callback}
\usage{
dt_callback(responsive = TRUE)
}
\arguments{
\item{responsive}{Boolean (default: \code{TRUE}). Whether to make table responsive or not}
}
\description{
Function to toggle responsive table on table created with \code{DT} package
}
