% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_box.R
\name{flip_box}
\alias{flip_box}
\title{flip_box}
\usage{
flip_box(
  id,
  front_content = h1("Hi From the Front"),
  back_content = tags$div(style =
    "background-color: #FFF; height: 380px; margin-top: -20px;", h1("Hi From the back"), )
)
}
\arguments{
\item{id}{the flipbox id}

\item{front_content}{ui for the front of the flip box}

\item{back_content}{ui for the back of the flip box}
}
\description{
ui for the flip_box_ui
}
