% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_box.R
\name{flip_button_back}
\alias{flip_button_back}
\title{flip_button_back}
\usage{
flip_button_back(id, text)
}
\arguments{
\item{id}{the flipbox id}

\item{text}{the button text}
}
\description{
button to flip to the back of the flip_box
}
\seealso{
Other flip_box functions: 
\code{\link{flip_box}()},
\code{\link{flip_button_front}()}
}
\concept{flip_box functions}
