% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-options.R
\name{hc_btn_options}
\alias{hc_btn_options}
\title{hc_btn_options}
\usage{
hc_btn_options()
}
\description{
options to only include specific download buttons in the highcharter download button
}
\examples{
library(highcharter)

# must be opened in a browser to work
hchart(rnorm(100)) \%>\%
  hc_exporting(
    enabled = TRUE,
    filename = "example-file",
    formAttributes = list(target = "_blank"),
    buttons = hc_btn_options()
  )

}
