% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_options.R
\name{hc_global_options}
\alias{hc_global_options}
\title{hc_global_options}
\usage{
hc_global_options()
}
\description{
Alters `highcharter` global options
}
\details{
Currently only change thousands separator from space to comma
}
