% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_button.R
\name{loading_button}
\alias{loading_button}
\title{loading_button}
\usage{
loading_button(
  id,
  label,
  style = "color: #FFFFFF; background-color: #337AB7; border-color: #2E6DA4;",
  loading_text = "Loading...",
  loading_style = "color: #FFFFFF; background-color: grey; border-color: grey;"
)
}
\arguments{
\item{id}{the flipbox id}

\item{label}{the button text (label)}

\item{style}{style for button (pre-loading); character string w/ CSS styling format: "color: black; background-color: red;"}

\item{loading_text}{text to show after button is clicked (e.g. during loading)}

\item{loading_style}{style for button (while loading); character string w/ CSS styling format: "color: black; background-color: red;"}
}
\description{
Button that becomes disabled until reset w/ `reset_loading_button`
}
\seealso{
Other flip_box functions: 
\code{\link{flip_box}()},
\code{\link{flip_button_back}()},
\code{\link{flip_button_front}()}
}
\concept{flip_box functions}
