% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_icon_cloud}
\alias{profile_icon_cloud}
\title{profile_icon_cloud}
\usage{
profile_icon_cloud(input, output, session, user = NULL)
}
\arguments{
\item{input}{no value needed for thos arguement}

\item{output}{no value needed for this arguement}

\item{session}{no value needed for this arguement}

\item{user}{optional reactive value representing user to override session$user}
}
\description{
server logic for placing profile icon in header bar of shinydashboard.
}
