% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_icon_cloud_ui}
\alias{profile_icon_cloud_ui}
\title{profile_icon_cloud_ui}
\usage{
profile_icon_cloud_ui(id)
}
\arguments{
\item{id}{character string identifying the module}
}
\description{
ui code for `profile_icon` for cloud.tychobra.com. Should be placed in the the `header`
argument to `dashboardPage()`
}
