% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_icon_ui}
\alias{profile_icon_ui}
\title{profile_icon_ui}
\usage{
profile_icon_ui(id)
}
\arguments{
\item{id}{character string identifying the module}
}
\description{
ui code for `profile_icon`. Should be placed in the the `header` arguement to
`dashboardPage()`
}
