% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_by.R
\name{read_by}
\alias{read_by}
\title{read_by}
\usage{
read_by(conn, tbl_name, cols = "*", by, operator = "AND")
}
\arguments{
\item{conn}{the DBI connection}

\item{tbl_name}{the name of the table}

\item{cols}{The columns to select from the table}

\item{by}{A named list of the field names and values to determine which rows to read}

\item{operator}{Either "AND" or "OR", determines whether the SQL query selects
the rows that match all of by or just at least one element of by}
}
\value{
the number of rows affected or an error. This value should be 1, 0, or the error
}
\description{
Get rows from a table by values in its columns
}
\examples{
con <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

DBI::dbWriteTable(
  con,
  name = "test",
  value = data.frame(
    id = 1:2,
    data_col = rep("hi", times = 2)
  )
)

dplyr::collect(dplyr::tbl(con, "test"))

read_by(con, "test", by = list(id = "1"))

DBI::dbDisconnect(con)



con <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

DBI::dbWriteTable(
  con,
  name = "test",
  value = data.frame(
    id = c(1,1,1,2,2),
    uid = c(1,2,1,1,2),
    data_col = rep("hi", times = 5)
  )
)

dplyr::collect(dplyr::tbl(con, "test"))

read_by(con, "test", by = list(id = "1", uid = "1"))

read_by(con, "test", by = list(id = "1", uid = "1"), operator = "Or")

read_by(con, "test", cols = "data_col", by = list(id = "1", uid = "1"))

read_by(con, "test", cols = c("uid", "data_col"), by = list(id = "1", uid = "1"))

DBI::dbDisconnect(con)
}
