% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_button.R
\name{reset_loading_button}
\alias{reset_loading_button}
\title{reset_loading_button}
\usage{
reset_loading_button(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{the input id}

\item{session}{the shiny session}
}
\description{
Reset the `loading_button` to its original style
}
