% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_apps.R
\name{run_loading_button_test_app}
\alias{run_loading_button_test_app}
\title{Run Loading Button Test App}
\usage{
run_loading_button_test_app()
}
\value{
shiny app
}
\description{
Run Loading Button Test App
}
\seealso{
loading_button
}
