% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_toast.R
\name{show_toast}
\alias{show_toast}
\title{show toast message}
\usage{
show_toast(type, message, .options = list())
}
\arguments{
\item{type}{length 1 character vector.  Valid values are "success", "error", "warning", and "info"}

\item{message}{the toast message}

\item{.options}{other options to pass to the `shinytoastr::toast_*()` function}
}
\value{
`invisible()`
}
\description{
A wrapper around the `shinytoastr::toast_*()` functions that uses our preferred default argument
values.
}
