% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboard-title.R
\name{tychobra_title}
\alias{tychobra_title}
\title{displays Tychobra logo next to shinydsahbord title}
\usage{
tychobra_title(title)
}
\arguments{
\item{title}{string - the title}
}
\description{
displays Tychobra logo next to shinydsahbord title
}
