% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by.R
\name{update_by}
\alias{update_by}
\title{update a row in a table}
\usage{
update_by(conn, tbl_name, by, .dat, operator = "AND")
}
\arguments{
\item{conn}{the DBI connection}

\item{tbl_name}{the name of the table}

\item{by}{A list of the field names and values to be updated}

\item{.dat}{a named list where the name is the database table field name and
each value is of length 1 and is the value to be updated}

\item{operator}{Either "AND" or "OR", determines whether the SQL query updates
all rows that match all of by or at least one element of by}
}
\value{
the number of rows affected or an error
}
\description{
update a row in a table
}
\examples{
con <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

DBI::dbWriteTable(
  con,
  name = "test",
  value = data.frame(
    id = 1:2,
    data_col = rep("hi", times = 2)
  )
)

test <- collect(tbl(con, "test")) 

test_dat <- list(
  data_col = "hello"
)

update_by(con, "test", by = list(id = 1), .dat = test_dat)

test <- collect(tbl(con, "test")) 

DBI::dbDisconnect(con)


con <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

DBI::dbWriteTable(
  con,
  name = "test",
  value = data.frame(
    id = c(1,2,1,1,2),
    uid = c(1,2,1,2,1),
    data_col = rep("hi", times = 5)
  )
)

test <- collect(tbl(con, "test")) 

test_dat <- list(
  data_col = "hello"
)

update_by(con, "test", by = list(id = 1, uid = 1), .dat = test_dat)

test <- collect(tbl(con, "test")) 

update_by(con, "test", by = list(id = 1, uid = 1), .dat = test_dat, operator = "or")

test <- collect(tbl(con, "test")) 

DBI::dbDisconnect(con)


}
