% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_by_id.R
\name{update_by_id}
\alias{update_by_id}
\title{update a row in a table}
\usage{
update_by_id(conn, tbl_name, id, .dat)
}
\arguments{
\item{conn}{the DBI connection}

\item{tbl_name}{the name of the table}

\item{id}{the row id}

\item{.dat}{a named list where the name is the database table field name and
each value is of length 1 and is the value to be updated}
}
\value{
the number of rows affected or an error
}
\description{
update a row in a table
}
\examples{
con <- DBI::dbConnect(
  RSQLite::SQLite(),
  dbname = ":memory:"
)

DBI::dbWriteTable(
  con,
  name = "test",
  value = data.frame(
    id = 1:2,
    data_col = rep("hi", times = 2)
  )
)

test_dat <- list(
  data_col = "hello"
)

update_by_id(con, "test", id = 1, .dat = test_dat)

DBI::dbDisconnect(con)

}
