% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_to_link.R
\name{url_to_link}
\alias{url_to_link}
\title{url_to_link}
\usage{
url_to_link(url, base_link = TRUE)
}
\arguments{
\item{url}{the string of the url}

\item{base_link}{boolean. If `TRUE` the basename of the url
will be displayed in the link.  If `FALSE`, the full url will
be displayed in the link}
}
\value{
the html for the link
}
\description{
convert a url to a link in html
}
\examples{
url_to_link("https://tychobra.com/howdy")

}
