% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tychobra_gh_labels.R
\name{use_tychobra_gh_labels}
\alias{use_tychobra_gh_labels}
\title{Use Tychobra Github Labels}
\usage{
use_tychobra_gh_labels(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[usethis:use_github_labels]{usethis::use_github_labels}}
  \describe{
    \item{\code{repo_spec}}{Optional repository specification (\code{owner/repo}) if you
don't want to target the current project.}
    \item{\code{labels}}{A character vector giving labels to add.}
    \item{\code{rename}}{A named vector with names giving old names and values giving
new names.}
    \item{\code{colours}}{Named character vectors giving hexadecimal
colours (like \code{e02a2a}) and longer descriptions. The names should match
label names, and anything unmatched will be left unchanged. If you
create a new label, and don't supply colours, it will be given a random
colour.}
    \item{\code{descriptions}}{Named character vectors giving hexadecimal
colours (like \code{e02a2a}) and longer descriptions. The names should match
label names, and anything unmatched will be left unchanged. If you
create a new label, and don't supply colours, it will be given a random
colour.}
    \item{\code{delete_default}}{If \code{TRUE}, removes GitHub default labels that do
not appear in the \code{labels} vector and that do not have associated issues.}
    \item{\code{auth_token}}{GitHub personal access token (PAT).}
    \item{\code{host}}{GitHub API host to use. Override with the endpoint-root for your
GitHub enterprise instance, for example,
"https://github.hostname.com/api/v3".}
  }}
}
\value{
invisibly returns \code{data.frame} of github labels
}
\description{
Apply Tychobra Specific Github Labels to a Repository.
}
