% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tychobra_js.R
\name{use_tychobra_js}
\alias{use_tychobra_js}
\title{use_tychobra_js}
\usage{
use_tychobra_js(show_window_icon = TRUE)
}
\arguments{
\item{show_window_icon}{boolean that deafults to `TRUE`.  Determines whether or
not to show Tychobra logo icon in browser tab}
}
\description{
function to load some custom js \code{shiny} and \code{shinydashboard} apps
}
