% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_module.R
\name{value_box_module}
\alias{value_box_module}
\title{value_box_module}
\usage{
value_box_module(input, output, server, value, subtitle)
}
\arguments{
\item{value}{The value to be displayed in the value box}

\item{subtitle}{The subtitle to be displayed in the value box}
}
\description{
value_box_module
}
