% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_module.R
\name{value_box_module_ui}
\alias{value_box_module_ui}
\title{value_box_module_ui}
\usage{
value_box_module_ui(id, icon = NULL, backgroundColor = "#7cb5ec",
  textColor = "#FFF", width = 4, href = NULL)
}
\arguments{
\item{icon}{An icon made by the shiny icon function}

\item{backgroundColor}{A hex color code string}

\item{textColor}{A hex color code string}

\item{width}{A number between 1 and 12}

\item{href}{A url}
}
\description{
value_box_module_ui
}
