% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxColor.R
\name{valueBoxColor}
\alias{valueBoxColor}
\title{Change the color of a value box}
\usage{
valueBoxColor(
  value,
  subtitle,
  icon = NULL,
  backgroundColor = "#7cb5ec",
  textColor = "#FFF",
  width = 4,
  href = NULL
)
}
\arguments{
\item{value}{The value to be displayed}

\item{subtitle}{The text below the value to be displayed}

\item{icon}{An icon tag created by icon}

\item{backgroundColor}{The background color of the value box}

\item{textColor}{The text color of the box}

\item{width}{The width of the box}

\item{href}{Option for a link}
}
\value{
HTML for a value box with a custom color
}
\description{
Change the color of a value box
}
\examples{
value <- "5\%"
subtitle <- "Credibility Percentage"
backgroundColor <- "#c6077d"

valueBoxColor(value, subtitle, backgroundColor = backgroundColor)
}
