% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_month_seq.R
\name{year_month_seq}
\alias{year_month_seq}
\title{generates sequence of "year_month" strings over a time period that includes all
months in the time period}
\usage{
year_month_seq(start_date, end_date)
}
\arguments{
\item{start_date}{the start_date for the time period}

\item{end_date}{the end date for the time period}
}
\value{
character vector with elements in the form "year_month". e.g.
c("2018-01", "2018-02")
}
\description{
generates sequence of "year_month" strings over a time period that includes all
months in the time period
}
\examples{
start <- as.Date("2018-01-07")
end <- as.Date("2018-03-01")

year_month_seq(start, end)

}
