% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_column_names.R
\name{convert_column_names}
\alias{convert_column_names}
\title{Convert values in a vector to their cooresponding values according to 
a data frame}
\usage{
convert_column_names(names_vector, names_map)
}
\arguments{
\item{names_map}{A data frame with columns "names" and "display names"}

\item{old_names}{The column names of a data frame}
}
\value{
The names where every old name in names is replaced with its cooresponding
display name
}
\description{
Convert values in a vector to their cooresponding values according to 
a data frame
}
\examples{
names_vector <- c("lot_location_code", "location_name", "actions")
names_map <- data.frame(
               names = c("location_name", "lot_location_code"),
               display_names = c("Location Name", "Lot Location Code"),
               stringsAsFactors = FALSE
               )
convert_column_names(names_vector, names_map)

}
