% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect.R
\name{db_connect}
\alias{db_connect}
\title{Database connection module}
\usage{
db_connect(db_config)
}
\arguments{
\item{db_config}{database configuration}
}
\value{
a DBIConnect object
}
\description{
connect to the db
}
