% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_app.R
\name{deploy_app}
\alias{deploy_app}
\title{Deploy Shiny application}
\usage{
deploy_app(
  deployment_type,
  deployed_dir_name,
  instance_name,
  project_id,
  project_zone,
  gcloud = NULL
)
}
\arguments{
\item{deployment_type}{Name of configuration in `config.yml` to use for deployed app}

\item{deployed_dir_name}{Name of the directory in the VM instance for this app}

\item{instance_name}{The name of the VM instance}

\item{project_id}{The name of the GCP project for this VM instance}

\item{project_zone}{The zone of the GCP project for this VM instance}

\item{gcloud}{Absolute path for gcloud CLI (Ex: `/usr/local/bin/google-cloud-sdk/bin`)}
}
\description{
Deploy a Shiny application to a GCP VM instance (Must have gcloud on local machine).
Currently, this function MUST be run with the application in a `shiny_app` directory, &
that directory should be in the Current Working Directory.
}
\examples{

\dontrun{
  deploy_app(
    deployment_type = 'production',
    deployed_dir_name = 'example_app',
    instance_name = 'instance-1',
    project_id = 'gcp-project-id',
    project_zone = 'us-east1-d',
    gcloud = '/usr/local/bin/google-cloud-sdk/bin'
  )
}


}
