% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_box.R
\name{flip_box}
\alias{flip_box}
\title{flip_box}
\usage{
flip_box(
  id,
  front_content = htmltools::h1("Hi From the Front"),
  back_content = tags$div(style = "background-color: #FFF; height: 380px;",
    back_content)
)
}
\arguments{
\item{id}{the flipbox id}

\item{front_content}{ui for the front of the flip box}

\item{back_content}{ui for the back of the flip box}
}
\description{
ui for the flip_box - based off of \href{https://rinterface.github.io/shinydashboardPlus/}{shinydashboardplus::flipBox}
}
\section{Warning}{
 this function is experimental.
}

\references{
\url{https://rinterface.github.io/shinydashboardPlus/reference/flipBox.html}
}
\seealso{
Other flip_box functions: 
\code{\link{flip_button_back}()},
\code{\link{flip_button_front}()}
}
\concept{flip_box functions}
