% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dt_time.R
\name{format_dt_time}
\alias{format_dt_time}
\title{format_dt_time}
\usage{
format_dt_time(col_num)
}
\arguments{
\item{col_num}{the column number to format.  `col_num` uses 1 based column indexing, so
the first column is column `1`.}
}
\description{
format_dt_time
}
\examples{
format_dt_time(5)

}
