% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_button.R
\name{loading_button}
\alias{loading_button}
\title{loading_button}
\usage{
loading_button(
  id,
  label,
  class = "btn",
  style = "color: #FFFFFF; background-color: #337AB7; border-color: #2E6DA4;",
  loading_label = "Loading...",
  loading_class = "btn",
  loading_style = "color: #FFFFFF; background-color: grey; border-color: grey;"
)
}
\arguments{
\item{id}{the input id}

\item{label}{the button text (label)}

\item{class}{the class(es) to apply to the button}

\item{style}{style for button (pre-loading); character string w/ CSS styling format: "color: black; background-color: red;"}

\item{loading_label}{text to show after button is clicked (e.g. during loading)}

\item{loading_class}{the loading button css class(es).}

\item{loading_style}{style for button (while loading); character string w/ CSS styling format: "color: black; background-color: red;"}
}
\description{
Button that becomes disabled until reset w/ `reset_loading_button`
}
