% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_icon}
\alias{profile_icon}
\title{profile_icon}
\usage{
profile_icon(input, output, session)
}
\arguments{
\item{input}{no value needed for thos arguement}

\item{output}{no value needed for this arguement}

\item{session}{no value needed for this arguement}
}
\description{
server logic for placing profile icon in header bar of shinydashboard.  Should be called without
any arguments
}
