% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tychobra_gh_labels.R
\name{use_tychobra_gh_labels}
\alias{use_tychobra_gh_labels}
\title{Use Tychobra Github Labels}
\usage{
use_tychobra_gh_labels()
}
\value{
invisibly returns \code{data.frame} of github labels
}
\description{
Apply Tychobra Specific Github Labels to a Repository.
}
